Fix ceilometer detection
The normal way to check if a service is enabled, is doing a
request for the endpoint, however, ceilometer api is deprecated
and there is no longer an endpoint available, so in this case
we must to query the services client and check if the metric
service is enabled or not. This patch do exactly that, and also
add the panko and gnocchi services in the list of services to be
checked because ceilometer tests depends on that two services as
well.

Change-Id: Id1923ad82be960691c8e8d8239432e0822b17690