Don't refetch same image over and over
During tempest configuration, we create two images - general and alt -
though they are backed by the same image file.  Each time we fetch the
same image from the internet, which is not optimal. This patch makes it
so that when the file already exists in the destination, it's taken on
face value and used to upload the image to the cloud.

Change-Id: I12600ea81ff2c634a720ab82c14433a434e548a1