Set scenario.img_file from the name of the image
After https://review.openstack.org/#/c/524156/, the name of the image
image saved locally does not match anymore the default value of
scenario.img_file, but the name of the downloaded file.
If those two values are different, as it is likely to happen every
time that a specific image is passed, the scenario tests and everything
else that use ScenarioTest.glance_image_create() fail.
Setting the property so that it matches the file name solves the issue.

Change-Id: I44936418ce359eaf04c4f47f0961687d65d45c6c