Fix discovery of identity versions
Currently when identity url is not a v2 address, python-tempestconf returns
an empty list of identity versions.
The patch tries to get all versions. If it fails it fails due to the fact that
the query for versions was version specific and therefor the response
contains only data about that version. At this point now the tool will return
a list containing only this version.
Also the patch adds a check if discovered versions' status is stable.

Change-Id: I2ce80eb9210e953586f1fe80deef21aae38c2731