Make tempestconf easier to use as an library
Make main and method responsible for parsing arguments more
modular, so that, it's easier to import them.

Now the argument parser can be imported to an other project,
we want to integrate to. This way, we achieve the CLI which
will be maintained at only one place.

Split main to two parts:
1. arguments handling
2. the rest of config_tempest, which can be importable
as an external module.

Change-Id: I3b1b9cbd05ccdb7ccae1b34211c4c955e831d32d