Fixing download of the cirros image
Download the image fails if the path doesn't exist.
img-dir in tempest.conf points to etc, so if you're running
python-tempestconf outside tempest dir, without tempest
overrides, it will point to default_dir/etc/cirros.
If etc doesn't exist, it will fail to download the image,
this patch fix it.

Change-Id: Ib44e57442da2aa656916327d70f8dccdc0c6d1be