Fix operator_role value for account-generator
When a tempest.conf is generated with admin credentials and
swift is enabled, object-storage.operator_role is set to admin.
This behavior is not wanted, when we generate tempest.conf
for tempest account-generator. In this case the operator_role
will be overriden in cli and set to Member (the same value
as tempest.conf with demo credentials contains)

Change-Id: I8b6e6a814a7cb96ef47bc6d3a8edc3f6a29aebfa