Fix check_service_status
The method returns None when discoverability is set explicitly
to True in CLI of python-tempestconf. The patch fixes that.

The patch also fixes one of the tests, where
self.Service.check_service_status was checked if it's True -
it's always True, it's a method.

A new unit test is added which sets discoverability simulating
CLI overrides and checks if check_service_status returns
an expected value.

Change-Id: I1690b5ff186bac28caf5113d771de8620d5a052c