Fix documentation structure
Usage section is there twice, it's because the README.rst
is included in doc too and it contains usage section as well.
The patch fixes that by not including the README.rst and
copies the content of README.rst to the documentation,
in order to have at least some, while documentation is being
written or reviewed.

Change-Id: I0247289029cd63b354965242ce7c1eca65f945eb
Story: 2002703
Task: 22543