Replace tripleo job
Replace tripleo-ci-centos-7-containers-multinode with
tripleo-ci-centos-7-scenario002-multinode-oooq-container due the fact
the last one has swift enabled and it's something that we want to test
since we are now supporting swift discovery in tempestconf.
Also notice the average time for this job is 2 hours and 35 minutes
while the previous one was 2 hours and 15 minutes, so we are adding 20
minutes, but worth since we are covering more tests.

Depends-On: https://review.openstack.org/#/c/580423/
Change-Id: I896dc49ff654f40bc2c7117d3edb884b86f635f9