Removing dependence from SERVICE_VERSIONS
This patch remove the dependence of some services to check for the
catalog and supported versions on constant SERVICE_VERSIONS. It also
remove this constant since it's no longer required.
In order to do that, four new methods were added in Service:
get_supported_versions, get_catalog, get_feature_name and
get_unversioned_service_name that
should be implemented by each service that requires this check.
Right now, the only services that uses are image, identity and volume.

Story: 2002787
Task: 22729
Change-Id: Ia41b7131dd1c1e76974ad4605566fd531572b974