Set use_dynamic_credentials with a priority
When an accounts.yaml file is used, python-tempestconf
has to set use_dynamic_credentials to False, otherwise
the generated tempest.conf doesn't work with Tempest.
However, values from a deployer input file are set with
a priority, so that they are not overriden during
discovery.

To fix the issue, use_dynamic_credentials has to be set
with priority too.

The related unit tests checks only a return value from
set function. The patch improves them, so that they
check if a value was overriden when it was supposed to
and the other way around.

Story: 2004140
Task: 27606
Depends-On: https://review.openstack.org/#/c/614312/
Change-Id: I76b8cf694116801e36587929bf320e5743534791