Fix diff in gates
The gates were using a diff command to determine if 2 tempest.conf
files contains the same sections, options and values.
However, this has turned out to be not a good approach, as the
values in tempest.conf can be written in a different order.
Therefore the patch adds a python script which is able to compare
values in 2 tempest.conf files regardless their order.

Change-Id: I65e30a6e3b4add39dd5b6eaf48e09ba450855dbc