Fix no _member_ role found
python-tempestconf sets _member_role by default, however,
the role is not created in devstack environment by default.
Therefore the patch adds logic which checks if the role is
present and if it's not, it tries to look for member role.
If the member role is not present as well, auth.tempest_roles
is not set at all.

Story: 2003550
Task: 24839

Change-Id: I0051bb4f870d58298403ac6fef6ae4ec454997af