Allow to specify min flavor memory and disk size
python-tempestconf doesn't provide any way how to define a
different flavor memory and disk size but provides an option
for defining a custom image (not cirros which is default).
However, when using custom image the default flavor size is
not enough.
Therefore the review adds two arguments which can be used
for defining flavor minimum memory and minimum disk size.

The review also edits the documentation and fixes a few typos.

Change-Id: Iba14b65ff8e147765e47d223dea6231ab649135f
Story: 2001536
Task: 6341