Remove heat_stack_owner role assigning to user
Prior to juno, and with earlier juno defaults, users needed to have
the heat_stack_owner role to use heat stack apis assign that role
to the user if the role is present.

Juno is very old release, deprecated and this functionality is no
longer needed so we are removing the code for the same.

Depends-On: https://review.opendev.org/#/c/660985/

Story: 2005756
Task: 33459

Change-Id: Ifd4f1c9162449864a91bdc480689679f902e1437
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>