Don't fail when heat-plugin is missing a value
Make sure that python-tempestconf will not fail due to a missing
value when heat-tempest-plugin is configured.
There has been a situation when python-tempestconf failed when it
was configuring [heat_plugin] section for a user whom intention
wasn't run the heat tests. Therefore the patch adds just a log
message there to inform about the missing value only users who
are going to run the heat-tempest-plugin.

Story: 2006277
Task: 35978

Change-Id: I2d10a03e444f29d3ffc4500e27be2e1688fa664b