Make name of the demo user unique
When tempestconf during the creation of a user discovered that the user
already existed then it expected that the password for this account has
a default value (secret). This could cause errors because the existing
account may have had a different password than the default value.

To prevent this from happening the tempestconf now creates a user named
'demo_tempestconf' instead of 'demo'. This change makes sure that collisions
are avoided. If the system already contains a user named 'demo_tempestconf'
then it was probably created by tempestconf.

Story: 2007665
Task: 39759
Change-Id: Iff8883b2d8fd80d417d5d5ee1bc9bebae15e213a