Make packstack jobs non voting
Considering that we are the only users of packstack job (based
on the http://codesearch.openstack.org/ search) which has been
failing for more than a week at the moment, this patch makes
the packstack jobs non voting.
The future of packstack jobs as gate ones within this project
will be discussed.

Change-Id: Idb059598f22fe51033cbd9ca18e596f7cc004fc0